add User schema, migration, and POST handler
