+++ package.json
@@ -12,6 +12,7 @@
     "@next-auth/prisma-adapter": "^1.0.7",
     "@prisma/client": "^6.6.0",
     "bcrypt": "^5.1.1",
+    "bcryptjs": "^3.0.2",
     "next": "15.2.4",
     "next-auth": "^4.24.11",
     "react": "^19.0.0",
