+++ src/app/createUser/page.tsx
@@ -1,0 +1,68 @@
+"use client";
+
+import { useState } from "react";
+
+export default function CreateUserPage() {
+  const [email, setEmail] = useState("");
+  const [username, setUsername] = useState("");
+  const [password, setPassword] = useState("");
+  const [message, setMessage] = useState("");
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+
+    const response = await fetch("/api/users", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify({ email, username, password }),
+    });
+
+    const data = await response.json();
+
+    if (response.ok) {
+      setMessage("User created successfully");
+      setEmail("");
+      setPassword("");
+    } else {
+      setMessage(data.error || "Something went wrong");
+    }
+  };
+
+  return (
+    <div className="p-8 max-w-md mx-auto">
+      <h1 className="text-2xl  mb-6 text-center">Create User</h1>
+      <form onSubmit={handleSubmit} className="flex flex-col gap-4">
+        <input
+          type="email"
+          placeholder="Email"
+          value={email}
+          onChange={(e) => setEmail(e.target.value)}
+          required
+          className="px-4 py-2 border-2"
+        />
+        <input
+          type="text"
+          placeholder="Username"
+          value={username}
+          onChange={(e) => setUsername(e.target.value)}
+          required
+          className="px-4 py-2 border-2"
+        />
+        <input
+          type="password"
+          placeholder="Password"
+          value={password}
+          onChange={(e) => setPassword(e.target.value)}
+          required
+          className="px-4 py-2 border-2"
+        />
+        <button type="submit" className="border-2 py-2">
+          Create Account
+        </button>
+      </form>
+      {message && <p className="mt-4 text-center text-sm ">{message}</p>}
+    </div>
+  );
+}
