+++ src/types/users.ts
@@ -1,0 +1,5 @@
+export interface CreateUserRequest {
+  email: string;
+  password: string;
+  username: string;
+}
